+++ src/PTDoc.Maui/MauiProgram.cs
@@ -62,6 +62,7 @@ public static class MauiProgram
         builder.Services.AddTransient<DraftAutosaveService>();
         builder.Services.AddScoped<IAdminApprovalService, AdminApprovalApiService>();
         builder.Services.AddScoped<INotificationCenterService, HttpNotificationCenterService>();
+        builder.Services.AddScoped<IDashboardAlertService, HttpDashboardAlertService>();
         builder.Services.AddScoped<IIntakeSessionStore, JsIntakeSessionStore>();
         builder.Services.AddScoped<IIntakeDemographicsValidationService, IntakeDemographicsValidationService>();
         builder.Services.AddScoped<IHeaderConfigurationService, HeaderConfigurationService>();